+++ board/safety/safety_hyundai_canfd.h
@@ -156,7 +156,7 @@ static void hyundai_canfd_rx_hook(CANPacket_t *to_push) {
   int bus = GET_BUS(to_push);
   int addr = GET_ADDR(to_push);
 
-  const int pt_bus = hyundai_canfd_hda2 ? 1 : 0;
+  const int pt_bus = hyundai_canfd_hda2 ? (int) 1 : (int) 0;
   const int scc_bus = hyundai_camera_scc ? 2 : pt_bus;
 
   if (bus == pt_bus) {
@@ -168,7 +168,7 @@ static void hyundai_canfd_rx_hook(CANPacket_t *to_push) {
     }
 
     // cruise buttons
-    const int button_addr = hyundai_canfd_alt_buttons ? 0x1aa : 0x1cf;
+    const int button_addr = hyundai_canfd_alt_buttons ? (int) 0x1aa : (int) 0x1cf;
     if (addr == button_addr) {
       int main_button = 0;
       int cruise_button = 0;
@@ -220,7 +220,7 @@ static void hyundai_canfd_rx_hook(CANPacket_t *to_push) {
   if (hyundai_longitudinal) {
     // on HDA2, ensure ADRV ECU is still knocked out
     // on others, ensure accel msg is blocked from camera
-    const int stock_scc_bus = hyundai_canfd_hda2 ? 1 : 0;
+    const int stock_scc_bus = hyundai_canfd_hda2 ? (int) 1 : (int) 0;
     stock_ecu_detected = stock_ecu_detected || ((addr == 0x1a0) && (bus == stock_scc_bus));
   }
   generic_rx_checks(stock_ecu_detected);
@@ -294,7 +294,7 @@ static int hyundai_canfd_fwd_hook(int bus_num, int addr) {
   }
   if (bus_num == 2) {
     // LKAS for HDA2, LFA for HDA1
-    int hda2_lfa_block_addr = hyundai_canfd_hda2_alt_steering ? 0x362 : 0x2a4;
+    int hda2_lfa_block_addr = hyundai_canfd_hda2_alt_steering ? (int) 0x362 : (int) 0x2a4;
     bool is_lkas_msg = ((addr == hyundai_canfd_hda2_get_lkas_addr()) || (addr == hda2_lfa_block_addr)) && hyundai_canfd_hda2;
     bool is_lfa_msg = ((addr == 0x12a) && !hyundai_canfd_hda2);
